Documentation
Added some more documentation to the code.
Moved DataSnapshot into its own py file.
