+++ website/app.py
@@ -3,20 +3,14 @@ import pandas as pd
 import zipfile, os, shutil, csv
 from zipfile import ZipFile
 from website.models import linear_regression as lr
+from website.models import snapshot as ds
+from werkzeug.utils import secure_filename
 
-"""Temp. Data Snapshot"""
-class DataSnapshot():
-    """This class handles keeping track of the data snapshot that the user submits."""
-    def __init__(self):
-        self.og_data = None
-        self.data = None
-        self.filename = None
 
-    def select_columns(self, x, y):
-        df = self.og_data[[x, y]].copy()
-        self.data = df
-
-snapshot = DataSnapshot()
+"""Flask Operation"""
+app = Flask(__name__)
+app.config["UPLOAD_FOLDER"] = "./temp/"
+snapshot = ds.snapshot()
 
 
 """Input Parsing Functions"""
@@ -30,22 +24,36 @@ def allowed_file(filename):
     return '.' in filename and type in allowed_extensions, ""
 
 
+def text_input_parse(s):
+    """This function takes in a string submitted by the user and converts it to a float or an integer. If an error occurs, it
+    returns an error to the user."""
+    try:
+        if '.' in s:
+            f = float(s)
+            format_f = "{:.{}f}".format(f, 3)
+            return float(format_f), ""
+        else:
+            return int(s), ""
+    except ValueError as e:
+        return ValueError, e
+
+
 def has_header(df):
     """Determine if the file has a header. It returns a list of headers to use when uploading the full file."""
     return isinstance(df.columns[0], str)
 
 
 def gen_headers(x):
+    """This function generate generic column names for a csv file based on the number of columns."""
     col = []
     for n in range(x): col.append("Column " + str(n+1))
     return col
 
 
+"""File Upload Functions"""
 def csv_upload(file):
     """This function takes a file input and converts it to a pandas DataFrame."""
     try:
-        print(app.config['UPLOAD_FOLDER'])
-        print(file)
         path = app.config['UPLOAD_FOLDER'] + file
         with open(path,"r") as f:
             sample = f.read(1024)
@@ -133,37 +141,18 @@ def zip_unpack(zip_file):
         return pd.concat(d, axis=0)
     except Exception as e:
         return "Program returned error while uploading the zip: " + str(e)
-
-
-def text_input_parse(s):
-    """This function takes in a string submitted by the user and converts it to a float or an integer. If an error occurs, it
-    returns an error to the user."""
-    try:
-        if '.' in s:
-            f = float(s)
-            format_f = "{:.{}f}".format(f, 3)
-            return float(format_f), ""
-        else:
-            return int(s), ""
-    except ValueError as e:
-        return ValueError, e
     
 
+"""Output Parsing Functions"""
 def get_graph_data(df):
     """Chart.js scatter plot requires the dataset to be in the format: {'x': , 'y': }."""
     json = df.copy().rename(columns={df.columns[0]: 'x', df.columns[1]: 'y'})
     return json
 
 
-def clean_data(df):
-    """This function handles cleaning the dataset. This is done by removing all null values from the set."""
-    df.dropna(inplace=True)
-
-def get_hypeparams(request):
-    #TODO: Fix html form w/ radio button inputs to make them toggable so a user can't choose
-    # choose both options by accident'
+"""Gather form information."""
+def get_hyperparams(request):
     val = []
-
     val.append(request.form['loss_strength'])
     val.append(request.form['penalty'])
     val.append(request.form['alpha'])
@@ -187,10 +176,6 @@ def get_hypeparams(request):
     return val
 
 
-"""Flask Operation"""
-app = Flask(__name__)
-app.config["UPLOAD_FOLDER"] = "./temp/"
-
 @app.route("/")
 def index():
     """Renders the home page of the website, the first page that a user will land on when visiting the website."""
@@ -202,7 +187,6 @@ def ml_form():
     """Renders the machine learning form for the linear regression model. This is done by pressing the button on the navigation bar."""
     if request.method == 'POST':
         # If step 1 of the ml_form has been completed, return new information
-        # Update this for WTForms later to better handle the data?
         if 'upload_file' in request.form:
             if 'file' not in request.files:
                 return render_template('ml_form.html',
@@ -216,51 +200,51 @@ def ml_form():
                             tab=0, 
                             filename=request.files['file'].filename,
                             error="No file submitted. Please submit a file with a valid extension (csv or zip).")
-            # Maybe add a checkbox to see if the data comes from UCI. They have a specific standards for uploading data we can use for uploading.
+            
+            
             bool, type = allowed_file(f.filename)
             if bool:
-                #f.save(f.filename)
-                #if zipfile.is_zipfile(f):
+                # If the uploaded file is a zip, send it to zip_unpack.
                 if type == 'zip':
-                    f.save(f.filename)
+                    f.save(secure_filename(f.filename))
                     result = zip_unpack(f.filename)
-                    if isinstance(result, str):
-                        return render_template('ml_form.html',
-                            tab=0, 
-                            filename=request.files['file'].filename,
-                            error=result)
-                    else:
-                        snapshot.og_data = result
 
+                # Else, the uploaded file must be a csv file, and should go to csv_upload
                 else:
                     f = request.files.get('file')
-                    f.save(os.path.join(app.config['UPLOAD_FOLDER'], f.filename))
+                    f.save(os.path.join(app.config['UPLOAD_FOLDER'], secure_filename(f.filename)))
                     result = csv_upload(f.filename)
 
-                    if isinstance(result, str):
-                        return render_template('ml_form.html',
-                            tab=0, 
-                            filename=request.files['file'].filename,
-                            error=result)
-                    else:
-                        snapshot.og_data = result
+                # If the upload functions return a string, an error was found and should be returned to the user.
+                if isinstance(result, str):
+                    return render_template('ml_form.html',
+                        tab=0, 
+                        filename=request.files['file'].filename,
+                        error=result)
+                
+                # Else, save the snapshot.
+                else:
+                    snapshot.og_data = result
 
+                # Return the output to the user.
                 return render_template('ml_form.html',
                             tab=0, 
                             file_upload=True,
                             filename=request.files['file'].filename,
                             og_df=snapshot.og_data.to_html(),
                             column_names=snapshot.og_data.columns.tolist())
+            
+            # In case something goes wrong, we ensure to render the template with a warning message.
             else:
                 return render_template('ml_form.html',
                             tab=0, 
                             filename=request.files['file'].filename,
                             error="Please submit a file with a valid extension (csv or zip).")
 
+        # If the user selects columns, display output.
+        # TODO: Added error handling if user same columns for X and Y
         if 'select_xy' in request.form:
             snapshot.select_columns(request.form['X'], request.form['Y'])
-
-
             df = get_graph_data(snapshot.data)
             return render_template('ml_form.html',
                            tab=0, 
@@ -273,9 +257,9 @@ def ml_form():
                            og_df=snapshot.og_data.to_html(),
                            column_names=snapshot.og_data.columns.tolist())
 
-        
+        # If the user submits the scaling form, clean the data and perform data scaling.
         if 'scaling' in request.form:
-            clean_data(snapshot.data)
+            snapshot.clean_data(snapshot.data)
             snapshot.data = lr.scaling(snapshot.data, request.form['scale'])
             return render_template('ml_form.html',
                            tab=1,
@@ -285,10 +269,12 @@ def ml_form():
                            titles=snapshot.data.columns.values,
                            og_df=snapshot.og_data.to_html())
         
+        # If the user submits the testing/training form
         if 'tt' in request.form:
-            print("Moving to Training / Testing Split")
             train, e = text_input_parse(request.form['training'])
             test, e = text_input_parse(request.form['testing'])
+            
+            # If the user submitted a non-integer/float value, return an error.
             if train is ValueError:
                 return render_template('ml_form.html',
                            tab=2,
@@ -304,12 +290,11 @@ def ml_form():
                            error=e,
                            og_df=snapshot.og_data.to_html())
 
-            df = snapshot.data
-            
-            #we need to generate colomn names so we can set x and y test/training splits
-            
-            train_df, test_df, msg = lr.test_train_split(df, test, train)
-
+            # df = snapshot.data
+            # Run the testing/train split.
+            train_df, test_df, msg = lr.test_train_split(snapshot.data, test, train)
+           
+            # If an error is found while trying to split the data, display the error.
             if train_df is None:
                 return render_template('ml_form.html',
                            tab=2,
@@ -317,27 +302,27 @@ def ml_form():
                            traintest=False,
                            error=msg,
                            og_df=snapshot.og_data.to_html())
+            
+            # Otherwise, get the json graph data and return the information needed for chartJS.
             test_df = get_graph_data(test_df)
             train_df=get_graph_data(train_df)
-
             return render_template('ml_form.html',
-                           tab=2,
-                           user_input=True,
-                           traintest=True,
-                           tr=train,
-                           te=test,
-                           og_df=snapshot.og_data.to_html(),
-                           test_name = "testing",
-                           training_name = "training",
-                           test_data=test_df.to_json(orient="records"),
-                           training_data=train_df.to_json(orient="records"),
-                           column_names=snapshot.data.columns.tolist(),
-                           error=msg
-                           )
+                                    tab=2,
+                                    user_input=True,
+                                    traintest=True,
+                                    tr=train,
+                                    te=test,
+                                    og_df=snapshot.og_data.to_html(),
+                                    test_name = "testing",
+                                    training_name = "training",
+                                    test_data=test_df.to_json(orient="records"),
+                                    training_data=train_df.to_json(orient="records"),
+                                    column_names=snapshot.data.columns.tolist(),
+                                    error=msg)
         
         if 'hyperparams' in request.form:
-            # TODO: Pull info from form. If text input, send to text_input_parse and return value.
-            val = get_hypeparams(request)
+            # TODO: Finish error handling here.
+            val = get_hyperparams(request)
             lr.initialize(val)
             return render_template('ml_form.html',
                            tab=3,
@@ -346,6 +331,7 @@ def ml_form():
                            og_df=snapshot.og_data.to_html())
         
         if 'run' in request.form:
+            # TODO: Implement fit + run
             return render_template('ml_form.html',
                            tab=4,
                            user_input=True,
